requirements.txt
aiohttp>=3.9.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
pyarrow>=14.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
